test(best-practices): unify all forms of test fixtures
All various forms of test case have been placed in `test/fixtures/testcases` in the same format; a directory containing a "testcases.json" file with the testcases for that particular BP rule, along with HTML files for those testcases.

I think this is the cleanest format, although it's a bit verbose in the directory.