test(util): tests for `isElementVisible()`
Several of the tests fail. These are the ones that work on the opacity of an element.

It looks like there's two things at play. First, a general misunderstanding of whether opacity 0 or 1 is fully transparent. 0 is fully transparent but the code treats it in the reverse. Second, it looks like the comparisons rely on `opacity && opacity === 0` to check that "opacity is defined and opacity is exactly 0" - but a quirk of comparisons instead evaluates this to `0` and not `false`. Looks like this bubbles up and might cause unintended results?